api/requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==3.0.10
pymongo==4.6.1
python-dotenv==1.0.0 